+++ learn-designpatterns-in-csharp/Source/Behavioral/Command/EmployeeManagerDemo/Commands/AddEmployeeToManagerList.cs
@@ -1,4 +1,5 @@
-﻿using EmployeeManagerDemo.Receivers;
+﻿using EmployeeManagerDemo.Models;
+using EmployeeManagerDemo.Receivers;
 
 namespace EmployeeManagerDemo.Commands
 {
