+++ learn-designpatterns-in-csharp/Source/Behavioral/Command/EmployeeManagerDemo/Commands/ICommand.cs
@@ -7,31 +7,10 @@
     public interface ICommand
     {
         void Execute();
+
         bool CanExecute();
+        
         void Undo();
-    }
-
-
-
-    public class Employee
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-
-        public Employee(int id, string name)
-        {
-            Id = id;
-            Name = name;
-        }
-    }
-
-    public class Manager : Employee
-    {
-        public List<Employee> Employees = new();
-        public Manager(int id, string name)
-            : base(id, name)
-        {
-        }
     }
 
 }
