+++ learn-designpatterns-in-csharp/Source/Behavioral/Command/EmployeeManagerDemo/EmployeeManagerDemo.csproj
@@ -7,4 +7,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Models\" />
+  </ItemGroup>
+
 </Project>
