+++ learn-designpatterns-in-csharp/Source/Behavioral/Command/EmployeeManagerDemo/Models/Employee.cs
@@ -1,0 +1,15 @@
+﻿namespace EmployeeManagerDemo.Models
+{
+    public class Employee
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+
+        public Employee(int id, string name)
+        {
+            Id = id;
+            Name = name;
+        }
+    }
+
+}
