+++ learn-designpatterns-in-csharp/Source/Behavioral/Command/EmployeeManagerDemo/Models/Manager.cs
@@ -1,0 +1,12 @@
+﻿namespace EmployeeManagerDemo.Models
+{
+    public class Manager : Employee
+    {
+        public List<Employee> Employees = new();
+        public Manager(int id, string name)
+            : base(id, name)
+        {
+        }
+    }
+
+}
