+++ learn-designpatterns-in-csharp/Source/Behavioral/Command/EmployeeManagerDemo/Receivers/EmployeeManagerRepository.cs
@@ -1,4 +1,4 @@
-﻿using EmployeeManagerDemo.Commands;
+﻿using EmployeeManagerDemo.Models;
 
 namespace EmployeeManagerDemo.Receivers
 {
