+++ learn-designpatterns-in-csharp/Source/Behavioral/Command/EmployeeManagerDemo/Receivers/IEmployeeManagerRepository.cs
@@ -1,4 +1,4 @@
-﻿using EmployeeManagerDemo.Commands;
+﻿using EmployeeManagerDemo.Models;
 
 namespace EmployeeManagerDemo.Receivers
 {
